Add Limine support (#835)
* Replace BootInfo by MemoryMap and offset

* Fix tests

* Add limine conf

* Add limine setup

* Add limine bios files

* Add limine image building to makefile

* Add limine crate

* Add limine start

* Set VGA text mode during init

* Add limine qemu to makefile

* Add debug to init

* Fix tests

* Add missing linker

* Use cargo-opts in makefile

* Add multiboot2 support

* Add option to compile to i686

* Fix json

* Move memory map to sys::boot module

* Split boot protocols into modules

* Refactor code

* Add comment

* Fix i686 compilation errors

* Remove init and exec calls from limine

* Fix compilation issues for i686

* Remove compilation warnings

* Move extract_memory_map to the bootloader module

* Fix tests

* Update makefile

* Move FileIO from api to sys

* Update limine conf

* Move imports

* Remove string allocation

* Add missing IO file

* Add MemoryMap::CAPACITY

* Add missing linker script

* Add missing step to limine setup

* Update makefile

* Replace xorriso with mtools

* Add grub config

* Fix sed command

* Fix x86 module to compile on i686

* Compile PIC on i686

* Compile x86::int on i686

* Enable serial output on i686

* Simplify makefile

* Add missing limine.ld

* Refactor makefile

* Add compile errors for incompatible features and targets

* Init serial on i686

* Remove dead code

* Avoid panic when adding memory region

* Move tmp/boot to run/boot

* Unify image name

* Add images to make clean

* Avoid hardcoding release mode

* Remove QEMU warning

* Fix typo in Makefile

* Remove grub support

* Update changelog

* Remove Grub from the Makefile

+++ src/sys/idt.rs
@@ -110,7 +110,7 @@ extern "x86-interrupt" fn page_fault_handler(
 ) {
     let csi_color = api::console::Style::color("red");
     let csi_reset = api::console::Style::reset();
-    let addr = Cr2::read();
+    let addr = Cr2::read() as u64;
     //debug!("EXCEPTION: PAGE FAULT ({:?}) at {:#X}", error_code, addr);
 
     let mut mapper = unsafe {
